+++ dev/imports-dev.php
@@ -31,6 +31,7 @@ function create_block_imports_dev_block_init() {
 	register_block_type( __DIR__ . '/build/blocks/urd-post-via-term' );
 	register_block_type( __DIR__ . '/build/blocks/urd-post-terms' );
 	register_block_type( __DIR__ . '/build/blocks/urd-post-meta' );
+	register_block_type( __DIR__ . '/build/blocks/post-chooser' );
 }
 add_action( 'init', 'create_block_imports_dev_block_init' );